@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface UserRepository extends MongoRepository<User, String> {
     Optional<User> findById(String s);
-    Optional<User> findByName(String s);
+    Optional<User> findByEmail(String s);
 }
